auth0/backend/requirements.txt
attrs==25.1.0
Authlib==1.5.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
ecdsa==0.19.0
flasgger==0.9.7.1
Flask==3.1.0
flask-cors==5.0.1
flask-swagger-ui==4.11.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwt==1.3.1
MarkupSafe==3.0.2
mistune==3.1.2
packaging==24.2
pyasn1==0.4.8
pycparser==2.22
python-dotenv==1.0.1
python-jose==3.4.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
rsa==4.9
six==1.17.0
urllib3==2.3.0
Werkzeug==3.1.3
